+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### pack
 
+* Added `@daltontan/postcss-import-json` PostCSS plugin to process JSON imports in CSS.
 * Fixed `eslintWebpackPluginConfig` to not always run lint in strict mode when `enact pack`.
 
 ## 7.0.0 (June 10, 2025)